requirements.txt
requests
configparser
pytesseract
selenium
pillow
numpy